Move stuff around.
sendCommand now takes an (optional) flags parameter, so it can be used on
DS and 3DS even if it can't be used w/ powersaves.
This should break existing usages slightly, so be aware.
This commit *shouldn't* change any behavior, as I don't have all the flags
on hand. Future commits can add the actual flags that should be there in.

format*Command -> send*Command, meaning API surrounding that is more ergonomic.

Update the README 'a bit'.